+++ lib/hooks/useScene/useSession.ts
@@ -13,7 +13,7 @@ import { Confetti } from "../../domains/confetti/Confetti";
 import { getUnix } from "../../domains/dayjs/getDayJS";
 import { IDiceRollResult } from "../../domains/dice/Dice";
 import { Id } from "../../domains/Id/Id";
-import { BlankTDDocument } from "../../routes/Draw/TLDraw";
+import { makeNewBlankDocument } from "../../routes/Draw/TldrawWriterAndReader";
 import { IPlayer, ISession } from "./IScene";
 import { IProps } from "./useScene";
 
@@ -36,7 +36,7 @@ export function useSession(props: IProps) {
       goodConfetti: 0,
       badConfetti: 0,
       paused: false,
-      tlDrawDoc: BlankTDDocument,
+      tlDrawDoc: makeNewBlankDocument(),
     })
   );
 
@@ -379,7 +379,7 @@ export function useSession(props: IProps) {
           return;
         }
         const everyone = getEveryone(draft);
-        draft.tlDrawDoc = BlankTDDocument;
+        draft.tlDrawDoc = makeNewBlankDocument();
         everyone.forEach((p) => {
           p.playedDuringTurn = false;
         });
