+++ package.json
@@ -102,8 +102,8 @@
     "@types/uuid": "^8.3.0",
     "@typescript-eslint/eslint-plugin": "^4.30.0",
     "@typescript-eslint/parser": "^4.30.0",
-    "@vitejs/plugin-legacy": "^1.6.2",
-    "@vitejs/plugin-react": "^1.0.9",
+    "@vitejs/plugin-legacy": "^1.6.4",
+    "@vitejs/plugin-react": "^1.1.3",
     "babel-loader": "^8.2.2",
     "chromatic": "^5.8.3",
     "concurrently": "^6.2.0",
@@ -120,6 +120,6 @@
     "react-test-renderer": "^17.0.2",
     "ts-jest": "^27.0.3",
     "typescript": "^4.4.2",
-    "vite": "^2.6.14"
+    "vite": "^2.7.10"
   }
 }
